First commit of DesktopController.
DesktopController is a static library when compiled in Debug or
Release. This can be referenced from other C++ projects.
DesktopController may also be compiled with python bindings
(using pybind11) which produces a shared library that can be
imported in python.

Two samples are included, DesktopSnake and ListIcons. The former
implements a game of snake which you can play using your desktop
icons. The latter demonstrated the various ways you can access
icons and list their details. A project containing fmtlib is also
present for use by the samples, but this is not part of the
DesktopController library.